Move babel-polyfill require to the CLI.
The Babel Polyfill is a giant pain, it only allows for one require call.
This causes issues when multiple entry points into an app exist, so it
was moved as far up the stack as possible to a CLI flag in order to
ensure that we don't have any clashes.